requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
jinja2==3.1.3
python-multipart==0.0.6
gspread==5.12.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dateutil==2.8.2
python-telegram-bot==22.6
